refactor(engines): split pulid_flux2_klein.py into _pulid/ siblings (#56)
Move Klein-specific PuLID nn.Modules to _pulid/klein_modules.py and
transformer patching + Strategy B dim-projection wiring to
_pulid/klein_patching.py. Engine file drops from 502 to 219 LOC.

- _pulid/klein_modules.py — PuLIDFlux2 + private helpers
- _pulid/klein_patching.py — patch_flux2 + _make_projections; carries
  the Strategy B rationale docstring
- _pulid/__init__.py — re-exports alongside existing flux1 symbols
- pulid_flux2_klein.py — engine + _extract_id_tokens only
- pulid_flux2_klein_fp4.py — updated to new imports
- tools/file_exemptions.txt — removes pulid_flux2_klein.py line
- tests/test_pulid_flux2_klein_init_order.py — CPU guard for
  Strategy B invariant (proj_up/proj_down must remain ephemeral,
  never attached to PuLIDFlux2 as sub-modules)

Closes #56
