+++ src/imagecli/engines/_pulid/__init__.py
@@ -1,13 +1,15 @@
-"""PuLID sub-package for the FLUX.1-dev face-lock engine.
+"""PuLID sub-package for the FLUX PuLID face-lock engines.
 
-Extracted from ``pulid_flux1_dev.py`` in #55 to keep the engine file under the
-300 LOC quality gate. Only the symbols re-exported here are considered public
-within ``imagecli``; helper functions in ``modules`` keep their leading
-underscore.
+Extracted from the monolithic engine files in #55 (FLUX.1-dev) and #56
+(FLUX.2-klein) to keep engine files under the 300 LOC quality gate. Only
+the symbols re-exported here are considered public within ``imagecli``;
+helper functions in the submodules keep their leading underscore.
 """
 
 from __future__ import annotations
 
+from .klein_modules import PuLIDFlux2
+from .klein_patching import patch_flux2
 from .modules import IDFormer, PerceiverAttention, PerceiverAttentionCA, PuLIDFlux1
 from .patching import patch_flux1
 
@@ -16,5 +18,7 @@ __all__ = [
     "PerceiverAttention",
     "PerceiverAttentionCA",
     "PuLIDFlux1",
+    "PuLIDFlux2",
     "patch_flux1",
+    "patch_flux2",
 ]
